Tests: share queue/browser handling for all worker types
- one queue to rule them all: browserstack and selenium
- retries and hard retries are now supported in selenium
- selenium tests now re-use browsers in the same way as browserstack

Close gh-508
